app/core/dependencies.py
# app/core/dependencies.py

from fastapi import Depends, HTTPException, status
from fastapi.security import OAuth2PasswordBearer
from jose import JWTError, jwt
from app.core.config import settings
from app.models.user import User
from app.db.mongodb import db

oauth2_scheme = OAuth2PasswordBearer(tokenUrl="api/v1/auth/login")
db_name = settings.PROJECT_NAME

async def get_current_user(token: str = Depends(oauth2_scheme)):
    credentials_exception = HTTPException(
        status_code=status.HTTP_401_UNAUTHORIZED,
        detail="Could not validate credentials",
        headers={"WWW-Authenticate": "Bearer"},
    )
    try:
        payload = jwt.decode(token, settings.SECRET_KEY, algorithms=["HS256"])
        email: str = payload.get("sub")
        if email is None:
            raise credentials_exception
    except JWTError:
        raise credentials_exception
    user_data = await db.client[db_name].users.find_one({"email": email})
    if user_data is None:
        raise credentials_exception
    user = User(**user_data)
    return user
